+++ NEWS.md
@@ -8,6 +8,7 @@
 ## Minor changes
    
 - New argument `show.legend.text` in the `ggscatter()` function. Use show.legend.text = FALSE to hide text in the legend.
+- New arguments `submain, caption, font.main, font.caption` in the `ggpar()` function.
    
 ## Bug fixed